Create a bundle for visualizer
Set up the config and code to create a bundle for visualizer and import
in WWW repo.
I separated the layout configs so that we only include the config that
we actully use.
