+++ packages/hyperion-autologging-visualizer/src/component/CytoscapeLayoutConfigDagre.ts
@@ -1,0 +1,48 @@
+/**
+ * Copyright (c) Meta Platforms, Inc. and affiliates. All Rights Reserved.
+ */
+
+import type cytoscape from 'cytoscape';
+import Cytoscape from 'cytoscape';
+import dagre from 'cytoscape-dagre';
+
+const DAGRE_CONFIG: dagre.DagreLayoutOptions = {
+  name: 'dagre',
+  // dagre algo options, uses default value on undefined
+  nodeSep: undefined, // the separation between adjacent nodes in the same rank
+  edgeSep: undefined, // the separation between adjacent edges in the same rank
+  rankSep: undefined, // the separation between each rank in the layout
+  rankDir: undefined, // 'TB' for top to bottom flow, 'LR' for left to right,
+  // align: undefined,  // alignment for rank nodes. Can be 'UL', 'UR', 'DL', or 'DR', where U = up, D = down, L = left, and R = right
+  // acyclicer: undefined, // If set to 'greedy', uses a greedy heuristic for finding a feedback arc set for a graph.
+  // A feedback arc set is a set of edges that can be removed to make a graph acyclic.
+  ranker: undefined, // Type of algorithm to assign a rank to each node in the input graph. Possible values: 'network-simplex', 'tight-tree' or 'longest-path'
+  minLen: function (_edge: any) { return 1; }, // number of ranks to keep between the source and target of the edge
+  edgeWeight: function (_edge: any) { return 1; }, // higher weight edges are generally made shorter and straighter than lower weight edges
+
+  // general layout options
+  fit: true, // whether to fit to viewport
+  padding: 30, // fit padding
+  spacingFactor: undefined, // Applies a multiplicative factor (>0) to expand or compress the overall area that the nodes take up
+  nodeDimensionsIncludeLabels: false, // whether labels should be included in determining the space used by a node
+  animate: true, // whether to transition the node positions
+  animateFilter: function (_node: any, _i: number) { return true; }, // whether to animate specific nodes when animation is on; non-animated nodes immediately go to their final positions
+  animationDuration: 500, // duration of animation in ms if enabled
+  animationEasing: undefined, // easing of animation if enabled
+  boundingBox: undefined, // constrain layout bounds; { x1, y1, x2, y2 } or { x1, y1, w, h }
+  transform: function (_node: any, pos: any) { return pos; }, // a function that applies a transform to the final node position
+  ready: function () { }, // on layoutready
+  sort: undefined, // a sorting function to order the nodes and edges; e.g. function(a, b){ return a.data('weight') - b.data('weight') }
+  // because cytoscape dagre creates a directed graph, and directed graphs use the node order as a tie breaker when
+  // defining the topology of a graph, this sort function can help ensure the correct order of the nodes/edges.
+  // this feature is most useful when adding and removing the same nodes and edges multiple times in a graph.
+  stop: function () { } // on layoutstop
+} as const;
+
+export function getCytoscapeLayoutConfig(): cytoscape.LayoutOptions {
+  const config = DAGRE_CONFIG;
+  console.log(`using ${config.name} layout config`);
+  Cytoscape.use(dagre);
+  return config;
+}
+
